moving sample source code about
