--- projenrc/nextjs/files/source/app/about/index.ts
@@ -1,1 +1,0 @@
-export * from "./page";
